+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - Fixed Incubator and Advanced Incubator from not working (hasRecipe() returned false hasEnoughEnergy was true rather than when false)
 - The bubbles in the Incubator and Advanced Incubator now animate slower
 - Fixed broken tooltip for Basic Genes in the Plasmid Infuser
+- Changed Plasmid tooltip text color to Gray